Move FOUND_PKG variables from constants file
