+++ model_compression_toolkit/exporter/model_wrapper/keras/validate_layer.py
@@ -15,7 +15,7 @@
 from typing import Any
 
 from mct_quantizers import BaseInferableQuantizer, KerasActivationQuantizationHolder
-from model_compression_toolkit.constants import FOUND_TF
+from model_compression_toolkit.verify_packages import FOUND_TF
 from model_compression_toolkit.logger import Logger
 
 if FOUND_TF:
