+++ model_compression_toolkit/gptq/pytorch/quantizer/base_pytorch_gptq_quantizer.py
@@ -16,7 +16,7 @@ from abc import abstractmethod
 from typing import Union, Dict
 
 from model_compression_toolkit.logger import Logger
-from model_compression_toolkit.constants import FOUND_TORCH
+from model_compression_toolkit.verify_packages import FOUND_TORCH
 from model_compression_toolkit.gptq.common.gptq_constants import WEIGHTS_QUANTIZATION_PARAMS
 
 from model_compression_toolkit.trainable_infrastructure import TrainableQuantizerWeightsConfig, \
