+++ model_compression_toolkit/pruning/pytorch/pruning_facade.py
@@ -15,7 +15,8 @@
 
 from typing import Callable, Tuple
 from model_compression_toolkit import get_target_platform_capabilities
-from model_compression_toolkit.constants import FOUND_TORCH, PYTORCH
+from model_compression_toolkit.constants import PYTORCH
+from model_compression_toolkit.verify_packages import FOUND_TORCH
 from model_compression_toolkit.core.common.mixed_precision.resource_utilization_tools.resource_utilization import ResourceUtilization
 from model_compression_toolkit.core.common.pruning.pruner import Pruner
 from model_compression_toolkit.core.common.pruning.pruning_config import PruningConfig
