+++ model_compression_toolkit/target_platform_capabilities/tpc_models/imx500_tpc/v1/tpc_keras.py
@@ -16,7 +16,7 @@ import tensorflow as tf
 from packaging import version
 
 from model_compression_toolkit.defaultdict import DefaultDict
-from model_compression_toolkit.constants import FOUND_SONY_CUSTOM_LAYERS
+from model_compression_toolkit.verify_packages import FOUND_SONY_CUSTOM_LAYERS
 from model_compression_toolkit.target_platform_capabilities.constants import KERNEL_ATTR, KERAS_DEPTHWISE_KERNEL, \
     KERAS_KERNEL, BIAS_ATTR, BIAS
