+++ model_compression_toolkit/trainable_infrastructure/keras/quantize_wrapper.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 from typing import Dict
 
-from model_compression_toolkit.constants import FOUND_TF
+from model_compression_toolkit.verify_packages import FOUND_TF
 from model_compression_toolkit.logger import Logger
 from model_compression_toolkit.trainable_infrastructure.common.base_trainable_quantizer import BaseTrainableQuantizer
